get score properly from DB in results

+++ GameEngine.ViewModel/ResultViewModel.cs
@@ -4,6 +4,6 @@
     {
         public GameViewModel GameViewModel { get; set; }
 
-        public decimal Score { get; set; }
+        public int Score { get; set; }
     }
 }
